requirements.txt
langchain
wikipedia
langchain-community
langchain-groq
langchain-anthropic
python-dotenv
pydantic
langchain-serpapi 
duckduckgo-search

